Crate transaction_engine

This package provides transaction processing for the toy transaction engine command-line tool, or for any other crate that might wish to use it.

In a production system we could imagine that this crate was in use by sofware running across a fleet of servers, with each server handling transactions for individual shards of users. This would work quite straightforward for the current toy spec as the types of transactions we are concerned about are per individual user and never between one user and the other. In theory then, if our toy code was to be deployed to a fleet of servers, we could shard on client id.

Beyond the toy spec though, something that would become a concern over time is balancing users across the fleet of servers. Certainly there is a lot that would need to be considered.

Going even further beyond the toy spec, there would certainly need to be all kinds of interaction between the transaction processing and other systems, and things would quickly become much more complex.

Examples of using this module

The following shows some basic examples of processing transactions using this module.

These examples also run as doc tests with cargo test --workspace when said command is executed in the toy transaction engine repository root.

In these examples, unwrap is used in order to keep the code short and also because we want the doc tests to fail except where the point is to show that invalid transactions are not accepted.

Deposits and withdrawals

use transaction_engine::{TransactionProcessor, ClientId, TransactionId, Accounts};

let mut transaction_processor = TransactionProcessor::new();

let client_a = ClientId::from(1u16);
let amount_1 = "1.5000000000001".try_into().unwrap();
let tx_1 = TransactionId::from(1u32);
let tx_2 = TransactionId::from(2u32);
// XXX: Precision of fractional part is four digits as per spec.
//      Fifth digit and onwards will be "chopped off", without rounding.
//
//      In other words, the following are all equivalent:
//
//let amount_2 = "0.25".try_into().unwrap();
//let amount_2 = "0.2500".try_into().unwrap();
//let amount_2 = "0.250099999999999999999999".try_into().unwrap();
let amount_2 = "0.25009".try_into().unwrap();

transaction_processor.deposit(client_a, tx_1, amount_1).unwrap();
transaction_processor.withdraw(client_a, tx_2, amount_2).unwrap();

// After processing the above two transactions, we expect to find that
// there is record of a single account, belonging to client_a. We expect that
// this account is not frozen, that the amount available on the account
// should have string representation equal to "1.2500", and that the held
// amount on the account should have string representation
// equal to "0.0000".
let accounts: Accounts = transaction_processor.into();
assert_eq!(accounts.len(), 1);
let (client_id_ret, acc_a) = accounts.into_iter().next().unwrap();
assert_eq!(client_id_ret, client_a);
assert!(!acc_a.is_frozen());
assert_eq!(acc_a.get_available().to_string(), "1.2500");
assert_eq!(acc_a.get_held().to_string(), "0.0000");

Dispute

use transaction_engine::{TransactionProcessor, ClientId, TransactionId, Accounts};

let mut transaction_processor = TransactionProcessor::new();

let client_a = ClientId::from(1u16);
let amount_1 = "1.5".try_into().unwrap();
let tx_1 = TransactionId::from(1u32);
let tx_2 = TransactionId::from(2u32);
let amount_2 = "0.25".try_into().unwrap();

transaction_processor.deposit(client_a, tx_1, amount_1).unwrap();
transaction_processor.withdraw(client_a, tx_2, amount_2).unwrap();
transaction_processor.dispute(client_a, tx_1).unwrap();

let accounts: Accounts = transaction_processor.into();
let (_, acc_a) = accounts.into_iter().next().unwrap();
assert_eq!(acc_a.get_available().to_string(), "-0.2500");
assert_eq!(acc_a.get_held().to_string(), "1.5000");

Resolve

use transaction_engine::{TransactionProcessor, ClientId, TransactionId, Accounts};

let mut transaction_processor = TransactionProcessor::new();

let client_a = ClientId::from(1u16);
let amount_1 = "1.5".try_into().unwrap();
let tx_1 = TransactionId::from(1u32);
let tx_2 = TransactionId::from(2u32);
let amount_2 = "0.25".try_into().unwrap();

transaction_processor.deposit(client_a, tx_1, amount_1).unwrap();
transaction_processor.withdraw(client_a, tx_2, amount_2).unwrap();
transaction_processor.dispute(client_a, tx_1).unwrap();
transaction_processor.resolve(client_a, tx_1).unwrap();

let accounts: Accounts = transaction_processor.into();
let (_, acc_a) = accounts.into_iter().next().unwrap();
assert_eq!(acc_a.get_available().to_string(), "1.2500");
assert_eq!(acc_a.get_held().to_string(), "0.0000");

Chargeback

use transaction_engine::{TransactionProcessor, ClientId, TransactionId, Accounts};

let mut transaction_processor = TransactionProcessor::new();

let client_a = ClientId::from(1u16);
let amount_1 = "1.5".try_into().unwrap();
let tx_1 = TransactionId::from(1u32);
let tx_2 = TransactionId::from(2u32);
let amount_2 = "0.25".try_into().unwrap();

transaction_processor.deposit(client_a, tx_1, amount_1).unwrap();
transaction_processor.withdraw(client_a, tx_2, amount_2).unwrap();
transaction_processor.dispute(client_a, tx_1).unwrap();
transaction_processor.chargeback(client_a, tx_1).unwrap();

let accounts: Accounts = transaction_processor.into();
let (_, acc_a) = accounts.into_iter().next().unwrap();
assert_eq!(acc_a.get_available().to_string(), "-0.2500");
assert_eq!(acc_a.get_held().to_string(), "0.0000");
assert_eq!(acc_a.get_total().to_string(), "-0.2500");
assert!(acc_a.is_frozen());

Structs

Account

Contains the account data for a single user.

ClientId

Client ID is represented by u16 integer as per spec.

FractionalAmount

Transaction amount is precise to four places past the decimal point in inputs and outputs. Therefore, we represent the amount internally as integer fractional amounts of 1/10,000ths (one ten thousands) of the i/o amount unit.

TransactionId

Transaction ID is represented by u32 integer as per spec.

TransactionProcessor

Processes transactions and provides final balances for accounts for which transactions have been processed.

Enums

FractionalAmountParseError

Errors returned for TryInto::try_into::<FractionalAmount> on &str.

TransactionChargebackError

Errors returned by TransactionProcessor::chargeback.

TransactionDepositError

Errors returned by TransactionProcessor::withdraw.

TransactionDisputeError

Errors returned by TransactionProcessor::dispute.

TransactionResolveError

Errors returned by TransactionProcessor::resolve.

TransactionWithdrawError

Errors returned by TransactionProcessor::withdraw.

Type Definitions

Accounts

Contains the accounts of all users for which we have processed valid transactions.